requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.36.0
bottle==0.12.25
bottle-websocket==0.2.9
Eel==0.16.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
macholib==1.16.2
Pillow==10.0.0
pygame==2.5.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.1.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
